- introduced signal busses  - cleaned up dsp API  - created classes bus, throw, catch  - created classes bus~, throw~, catch~

+++ packages/guiobj/c/src/display.c
@@ -188,7 +188,7 @@ display_ftl(fts_word_t *argv)
 
   for(i=0; i<n_tick; i++)
     {
-      float absval = fabs(in[i]);
+      float absval = fabs((double)in[i]);
 
       if(absval > this->absmax)
 	this->absmax = absval;
